Add logout and clear backstack

+++ libfoundation/src/main/java/com/fxapp/libfoundation/data/LatestRates.kt
@@ -1,7 +1,7 @@
 package com.fxapp.libfoundation.data
 
 data class LatestRates (
-    val amount: Long,
+    val amount: Double,
     val base: String,
     val date: String,
     val rates: Map<String, Double>
